App completo.

+++ ShoppingApp/Model/Produto.swift
@@ -12,6 +12,6 @@ struct Produto{
     let nomeDisco: String
     let nomeBanda: String
     let ano: String
-    let preco: Float
+    let preco: Double
     let capa: String
 }
